bring in hypotheses
