+++ server.py
@@ -1,15 +1,36 @@
 import socket
-from network_utils import get_host_name, PORT, FORMAT
+import network_utils as utils
+from network_utils import PORT, FORMAT
 
-SERVER = get_host_name()
+SERVER = utils.get_host_name()
 
 server = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
 server.bind((SERVER, PORT))
 
+PLAYER_DATA = {}
+
 
 def respond(data, addr):
-    msg = f'you said "{data.decode(FORMAT)}"'
-    server.sendto(msg.encode(FORMAT), addr)
+    player_id, x, y = utils.decode_network_data(data)
+
+    print(f"Player {player_id} is at position {x}, {y}")
+    PLAYER_DATA[player_id] = {
+        "data": utils.encode_network_data(player_id, x, y),
+        "x": x,
+        "y": y,
+    }
+
+    # TODO: how to keep track of opponent? receive id?
+    # Find a player id that's not the one just received
+    pids = PLAYER_DATA.keys()
+
+    try:
+        other_player_id = [pid for pid in pids if pid != player_id][0]
+        msg = PLAYER_DATA[other_player_id]["data"]
+    except IndexError:
+        msg = "No opponent".encode(FORMAT)
+
+    server.sendto(msg, addr)
 
 
 def start_server():
